pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distribution-strategy"
version = "0.2.0"
description = "Content scheduling, multi-channel distribution, and analytics for the marketing layer"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[tool.setuptools.packages.find]
include = ["kerygma_strategy*"]

[project.scripts]
distrib = "kerygma_strategy.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 99
